ninary cross

+++ scripts/utils.py
@@ -32,7 +32,8 @@ def sensory_attenuation(predicted_opt_flow, next_image, background_image, unnorm
 def binarize_optical_flow(param, optflow, positive_value = 255):
     #if param.get('opt_flow_binarize'):
     #    return np.array(np.where(optflow > param.get('opt_flow_binary_threshold'), positive_value, 0), dtype='uint8')
-    return np.array(optflow, dtype='uint8')
+    #return np.array(optflow, dtype='uint8')
+    return np.array(optflow)
 
 def intersection_over_union(param, y_true, y_pred, is_true_binarised, is_pred_binarised):
     #if not is_true_binarised:
